rename and add home btn
