+++ src/index.css
@@ -6,6 +6,18 @@
   font-family: 'Playfair Display', serif;
 }
 
+header {
+  height: auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding-inline: 2rem;
+  border-bottom: 16px solid rgb(227, 7, 19);
+  height: 140px;
+  width: 100%;
+  background-color: rgb(255, 249, 236);
+}
+
 .logo {
   width: auto;
   height: 100px;
@@ -28,19 +40,6 @@
 html, body {
   margin: 0;
   padding: 0;
-  background-color: rgb(255, 249, 236);
-}
-
-
-header {
-  height: auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-inline: 2rem;
-  border-bottom: 16px solid rgb(227, 7, 19);
-  height: 140px;
-  width: 100%;
   background-color: rgb(255, 249, 236);
 }
