+++ src/styles/RecipeCard.css
@@ -1,5 +1,3 @@
-
-
 @keyframes fadeIn {
     from {
         opacity: 0;
@@ -58,13 +56,10 @@ h4 {
     padding: 0;
     margin: 0;
     margin: 0rem 1rem;
-
-}
-
-h4 {
     color: rgb(71, 71, 71);
     font-size: 0.8rem;
-    word-wrap: break-word; /* Add this line */
+    word-wrap: break-word;
+
 }
 
 .cardBottomText {
